+++ standalone/android-arm64-v8a/external.sh
@@ -2,7 +2,7 @@
 
 set -e
 
-FREEIMAGE_VERSION=3.18.0
+FREEIMAGE_SHA=3532c976f9853219d42d8cbab6c1f4c0990e3ed8
 SDL_SHA=b5c3eab6b447111d3c7879bb547b80fb4abd9063
 SDL_IMAGE_SHA=4a762bdfb7b43dae7a8a818567847881e49bdab4
 SDL_TTF_SHA=07e4d1241817f2c0f81749183fac5ec82d7bbd72
@@ -24,7 +24,7 @@ else
 fi
 
 echo "Building external libraries..."
-echo "  FREEIMAGE_VERSION: ${FREEIMAGE_VERSION}"
+echo "  FREEIMAGE_SHA: ${FREEIMAGE_SHA}"
 echo "  SDL_SHA: ${SDL_SHA}"
 echo "  SDL_IMAGE_SHA: ${SDL_IMAGE_SHA}"
 echo "  SDL_TTF_SHA: ${SDL_TTF_SHA}"
@@ -59,30 +59,33 @@ cd tmp
 # build freeimage, and copy to external
 #
 
-CACHE_NAME="FreeImage-${FREEIMAGE_VERSION}"
+CACHE_NAME="FreeImage-${FREEIMAGE_SHA}"
 
 if [ ! -f "../${CACHE_DIR}/${CACHE_NAME}.cache" ]; then
-   FREEIMAGE_BASENAME="FreeImage${FREEIMAGE_VERSION//./}"
-   curl -sL https://downloads.sourceforge.net/project/freeimage/Source%20Distribution/${FREEIMAGE_VERSION}/${FREEIMAGE_BASENAME}.zip -o ${FREEIMAGE_BASENAME}.zip
-   unzip ${FREEIMAGE_BASENAME}.zip
-   cd FreeImage
-   cp ../../freeimage/Android.mk Android.mk
-   $ANDROID_NDK_HOME/ndk-build \
-      NDK_PROJECT_PATH=$ANDROID_NDK_HOME \
-      APP_BUILD_SCRIPT=Android.mk \
-      APP_STL="c++_static" \
-      APP_PLATFORM=android-30 \
-      APP_ABI=arm64-v8a \
-      NDK_OUT=obj \
-      NDK_LIBS_OUT=libs \
-      -j${NUM_PROC}
+   curl -sL https://github.com/toxieainc/freeimage/archive/${FREEIMAGE_SHA}.zip -o freeimage-${FREEIMAGE_SHA}.zip
+   unzip freeimage-${FREEIMAGE_SHA}.zip
+   ls -laRt ../..
+   cd freeimage-${FREEIMAGE_SHA}
+   cp ../../freeimage/CMakeLists.txt .
+   cp ../../freeimage/PluginEXR.cpp Source/FreeImage/PluginEXR.cpp
+   cmake \
+      -DPLATFORM=android \
+      -DARCH=arm64-v8a \
+      -DBUILD_STATIC=OFF \
+      -DCMAKE_BUILD_TYPE=${BUILD_TYPE} \
+      -B build
+   cmake --build build -- -j${NUM_PROCS}
+   mkdir -p ../../${CACHE_DIR}/${CACHE_NAME}/include
+   cp Source/FreeImage.h ../../${CACHE_DIR}/${CACHE_NAME}/include
    mkdir -p ../../${CACHE_DIR}/${CACHE_NAME}/lib
-   cp libs/arm64-v8a/*.so ../../${CACHE_DIR}/${CACHE_NAME}/lib
+   cp build/*.so ../../${CACHE_DIR}/${CACHE_NAME}/lib
    cd ..
    touch "../${CACHE_DIR}/${CACHE_NAME}.cache"
 fi
 
+cp -r ../${CACHE_DIR}/${CACHE_NAME}/include/* ../external/include
 cp ../${CACHE_DIR}/${CACHE_NAME}/lib/*.so ../external/lib
+
 
 #
 # download bass24 and copy to external
@@ -225,7 +228,7 @@ cp ../${CACHE_DIR}/${CACHE_NAME}/lib/*.so ../external/lib
 
 CACHE_NAME="libaltsound-${LIBALTSOUND_SHA}"
 
-if [ ! -f  "../${CACHE_DIR}/${CACHE_NAME}.cache" ]; then
+if [ ! -f "../${CACHE_DIR}/${CACHE_NAME}.cache" ]; then
    curl -sL https://github.com/vpinball/libaltsound/archive/${LIBALTSOUND_SHA}.zip -o libaltsound.zip
    unzip libaltsound.zip
    cd libaltsound-$LIBALTSOUND_SHA